refactor SR saving feature
